Create BUILDEVENT_FILE during start step (#22)
The buildevents CLI populates its traces and spans by parsing the file at BUILDEVENT_FILE. Because we created this file at the end of the build, our GitHub Actions-specific fields are only present on the last two spans.
This change creates the file at the start of the build and appends fields to the file later. This way, every span contains the github.* fields.