Remove plural from API secret
This is a bit confusing and inconsistent. In particular, in the [baseline buildevent library README](https://github.com/honeycombio/buildevents#environment-variables), it is suggested to use `BUILDEVENT_APIKEY`. And that is consistent with the [CircleCI orb docs](https://github.com/honeycombio/buildevents-orb#overview).

If people are going between the two docs it is very possible that they will introduce a hard to identify bug.

NOTE: this requires updating of the new guide. If this PR is merged, I will take that on immediately following.